Include my username in manifest.json metadata
I don't know if this is something you want, but I figured I'd check with you. In the manifest, I added my username, since I contributed to this repository. I thought that might be a good form of credit for anyone who contributes. If you don't want this, I fully understand.